requirements.txt
# # --------- pytorch --------- #
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.0+cu117
torchvision==0.15.1+cu117

lightning[extra]==2.0.0
torchmetrics==0.11.4

# # # --------- model zoo -------#
transformers==4.30.2
timm==0.9.2

# # ------- tensor utils ------#
einops==0.6.1

# # --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# # --------- loggers --------- #
wandb==0.15.4
# mlflow==2.4.1  # install this later with pip install as it is failing with devcontainer build as its trying to uninstall blinker==1.4 and install 1.6 instead
aim==3.17.5     # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# # --------- deployment ------ #
fastapi==0.88.0

# # --------- others --------- #
pyrootutils==1.0.4      # standardizing the project root setup
pre-commit==3.3.3       # hooks for applying linters on commit
rich==13.4.2            # beautiful text formatting in terminal
pytest==7.3.2           # testing
black==23.3.0           # code formatting
twine==4.0.2             # for uploading to pypi
sh==2.0.4               # for running bash commands in some tests (linux/macos only)

# # --------- data version control --------- #
dvc==3.0.0              # data version control
dvc-gdrive==2.19.2      # google drive remote for dvc